requirements.txt
pyinstrument
python-jose
aiofiles
alembic==1.11.1
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==23.1.0
argon2-cffi-bindings==25.1.0
asyncpg==0.30.0
bcrypt==5.0.0
cffi==2.0.0
click==8.3.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.118.0
greenlet==3.2.4
h11==0.16.0
httptools==0.6.4
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.3
passlib==1.7.4
psycopg2-binary==2.9.10
pwdlib==0.2.1
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
PyJWT==2.10.1
python-multipart==0.0.20
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.37.0
google-genai==1.63.0
pydantic-settings
uliweb-alembic
celery==5.3.*
redis>=4.5
protobuf
kombu
pillow
PyMuPDF==1.27.1
numpy